Move css to separate file for barpage
